+++ CHANGELOG.md
@@ -122,7 +122,7 @@
 * Define `Transition` type alias
 * Define `Stateful` type alias
 * Define `Scenario` type alias
-* Define `Stateful` pattern synonym
+* Define `Statefully` pattern synonym
 * Define `Scenario` pattern synonym
 * Define `rewrap` expression
 * Define `/` operator